+++ app/src/main/java/org/simple/clinic/search/results/PatientSearchResultsScreenKey.kt
@@ -2,15 +2,14 @@ package org.simple.clinic.search.results
 
 import kotlinx.android.parcel.IgnoredOnParcel
 import kotlinx.android.parcel.Parcelize
-import org.simple.clinic.R
+import org.simple.clinic.navigation.v2.ScreenKey
 import org.simple.clinic.patient.PatientSearchCriteria
-import org.simple.clinic.router.screen.FullScreenKey
 
 @Parcelize
-data class PatientSearchResultsScreenKey(val criteria: PatientSearchCriteria) : FullScreenKey {
+data class PatientSearchResultsScreenKey(val criteria: PatientSearchCriteria) : ScreenKey() {
 
   @IgnoredOnParcel
   override val analyticsName = "Patient Search Results"
 
-  override fun layoutRes() = R.layout.screen_patient_search_results
+  override fun instantiateFragment() = PatientSearchResultsScreen()
 }