+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Migrate `TeleconsultSharePrescriptionScreen` to a fragment
 - Remove deprecated Mobius delegate usage in `BloodSugarHistoryScreen`
 - Bump MDC to v1.3.0
+- Change `PatientSearchResultsScreen` to a fragment
 - [In Progress: 20 Jan 2021] Material Theming Migration
 - [In Progress: 08 Feb 2021] Migrate app to use ViewBinding